AI/pyproject.toml
[tool.poetry]
name = "face"
version = "0.1.0"
description = ""
authors = ["c"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
cloudinary = "^1.41.0"
pre-commit = "^4.0.1"
pylint = "^3.3.2"
django = "^5.1.4"
djangorestframework = "^3.15.2"
torch = "^2.5.1"
torchvision = "^0.20.1"
pillow = "^11.0.0"
opencv-python = "^4.10.0.84"
numpy = "^2.2.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
